requirements.txt
pillow 
exifread<3
PyExifTool